Implement transcription manager
Add TranscriptionManager object with Transcribe script. Change
'SaveUserSpeech' to 'NPCManager' and update the code
accordingly. Move all transcription code from here to new
class.

Issue Number: #337
