+++ Assets/VR4VET/Components/NPC/AI/AIConversationController.cs
@@ -5,7 +5,9 @@ using UnityEngine;
 public class AIConversationController : MonoBehaviour
 {
 
-    [SerializeField] private SaveUserSpeech _SaveUserSpeech;
+    [SerializeField] private NPCManager _NPCManager;
+
+    [SerializeField] private Transcribe _Transcribe;
     
     [TextArea(3, 10)]
     public string contextPrompt;
@@ -14,21 +16,23 @@ public class AIConversationController : MonoBehaviour
     public List<Message> messages = new List<Message>();
 
     
-
     void Start()
     {
-        _SaveUserSpeech = GetComponent<SaveUserSpeech>();
+        _NPCManager = GetComponent<NPCManager>();
+        _Transcribe = GameObject.Find("TranscriptionManager").GetComponent<Transcribe>();
     }
 
     public void StartRecording()
     {
-        _SaveUserSpeech.StartRecording(contextPrompt, maxTokens);
+        // Passing current NPCManager object to the TranscriptionManager so that it correctly handles awaiting transcripts
+        _Transcribe.StartRecording(_NPCManager);
+        _NPCManager.StartRecording(contextPrompt, maxTokens);
     }
 
 
-    public void EndRecording()
+    public async void EndRecording()
     {
-        _SaveUserSpeech.EndRecording();
+        _Transcribe.EndRecording();
     }
 
     public void AddMessage(Message message)
