+++ Assets/VR4VET/Components/NPC/AI/NPCManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3d4c489dca070914ab275a92f436d413
+guid: 0ab2a19051d1daa4d81f173ac85c27cd
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
