+++ Assets/VR4VET/Components/NPC/AI/Transcribe.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 12023bcc27be0664da87729f3264d711
+guid: c1fa4a43fe3e1ea4ab231f8f9aea769f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
